refactor: Book 어노테이션 및 타입 변경
