+++ Readio/src/main/java/ninegle/Readio/book/domain/Publisher.java
@@ -21,15 +21,13 @@ import lombok.NoArgsConstructor;
  * purpose: 
  */
 @Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 public class Publisher {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private long id;
 
 	@Column(nullable = false)
 	private String name;
