+++ Readio/src/main/java/ninegle/Readio/book/repository/AuthorRepository.java
@@ -1,6 +1,7 @@
 package ninegle.Readio.book.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import ninegle.Readio.book.domain.Author;
 
@@ -11,5 +12,6 @@ import ninegle.Readio.book.domain.Author;
  * author:  gigol
  * purpose: 
  */
+@Repository
 public interface AuthorRepository extends JpaRepository<Author,Long> {
 }
