+++ Readio/src/main/java/ninegle/Readio/book/repository/BookSearchRepository.java
@@ -3,8 +3,9 @@ package ninegle.Readio.book.repository;
 import java.util.List;
 
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
+import org.springframework.stereotype.Repository;
 
-import ninegle.Readio.book.dto.BookEsDto;
+import ninegle.Readio.book.domain.BookSearch;
 
 /**
  * Readio - BookSearchRepository
@@ -13,6 +14,7 @@ import ninegle.Readio.book.dto.BookEsDto;
  * author:  gigol
  * purpose: 
  */
-public interface BookSearchRepository extends ElasticsearchRepository<BookEsDto,Long> {
-	List<BookEsDto> findByTitleContainingOrPublisherContainingOrAuthorContaining(String title,String publisher,String author);
+@Repository
+public interface BookSearchRepository extends ElasticsearchRepository<BookSearch,Long> {
+	List<BookSearch> findByTitleContainingOrPublisherContainingOrAuthorContaining(String title,String publisher,String author);
 }
